refactor: split chat providers into separate modules
